Split CMakeLists.txt in seperate files in src, docs. (#181)
* Split CMakeLists.txt in seperate files in src, docs.
Fix sid extract in out of source build

* Added cmakelist files

* Fix MPI build and pin coulombgalore to specific commit

* enable_sid=on by default; skip quietly if sdl2 not found
