+++ src/analysis.h
@@ -2,7 +2,6 @@
 
 #include "space.h"
 #include "io.h"
-#include "mpi.h"
 #include "scatter.h"
 #include "reactioncoordinate.h"
 #include "auxiliary.h"
