+++ src/core_test.h
@@ -1,7 +1,7 @@
 #include "core.h"
 #include "units.h"
 #include "particle.h"
-#include "src/aux/iteratorsupport.h"
+#include "aux/iteratorsupport.h"
 #include <range/v3/view/filter.hpp>
 
 namespace Faunus {
