+++ src/speciation.cpp
@@ -1,6 +1,6 @@
 #include "bonds.h"
 #include "speciation.h"
-#include "src/aux/iteratorsupport.h"
+#include "aux/iteratorsupport.h"
 
 namespace Faunus {
 namespace Move {
